backend/requirements.txt
PyPDF2 == 3.0.1 
pymupdf == 1.26.3
dotenv
google.generativeai == 0.8.5
flask == 3.1.2
flask
flask-cors